+++ examples/resources/meraki_camera_quality_and_retention/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_camera_quality_and_retention.example "<serial>"